git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@871 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ utils/delay/generate_delay_model.h
@@ -22,7 +22,7 @@ struct Station_data {
 
   // EOP information.
   double tai_utc; // s
-  long   eop_ref_epoch; // julian day
+  double eop_ref_epoch; // julian day
   int    num_eop_points;
   double ut1_utc[10];  // sec
   double x_wobble[10]; // asec
